feat: refactor, imports and try me outs
