+++ tests/endpoints/test_availability.py
@@ -1,5 +1,5 @@
 
-from vortexasdk.endpoints.availability_search import VesselAvailabilitySearch
+from vortexasdk import VesselAvailabilitySearch
 from tests.testcases import TestCaseUsingRealAPI
 
 rotterdam = "68faf65af1345067f11dc6723b8da32f00e304a6f33c000118fccd81947deb4e"
