+++ tests/endpoints/test_utilisation_speed_breakdown.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from vortexasdk.endpoints.utilisation_speed_breakdown import FleetUtilisationSpeedBreakdown
+from vortexasdk import FleetUtilisationSpeedBreakdown
 
 from docs.utils import to_markdown
 from tests.testcases import TestCaseUsingRealAPI
